requirements.txt
gunicorn
django-heroku
djangorestframework
markdown
django-filter
django-cors-headers
coreapi
djangorestframework_simplejwt